requirements.txt
# AI & Agents
google-adk
a2a-sdk==0.3.14
google-generativeai>=0.8.0
google-genai>=0.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
nest-asyncio>=1.6.0

# Database
sqlalchemy>=2.0.0

# HTTP & API
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.40.0

# PDF Processing
pymupdf>=1.24.0

# Stock Data
yfinance>=0.2.40

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
cachetools>=5.3.0

# FastAPI for A2A endpoints
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# UI
streamlit>=1.32.0

# Async support
asyncio-mqtt>=0.16.0
aiofiles>=23.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.8.0
types-requests>=2.31.0
types-beautifulsoup4>=4.12.0

# Utilities
python-dateutil>=2.8.0
pytz>=2024.1

# Observability
openinference-instrumentation-google-adk==0.1.6
arize-phoenix-otel>=0.14.0